Remove unnecessary spacing resulting from specter upgrade (#1692)
# What it does

The spectre.css upgrade in #1686 ended up messing up a bunch of spacing
in the app, mostly due to changes to the default styles of `p`, `ul`,
and `li` elements. I clicked around and fixed the most glaring visual
issues, but I wouldn't be surprised if there are more.

I also ended up fixing #893 in the process.