Tweak sign up for clarity around payments (#1699)
# What it does

Changes some copy and has the form require a minimum of a $5 online
donation

# Why it is important

Takes care of the requests in #1682 

# UI Change Screenshot

New copy:
![Screenshot 2024-10-07 at 7 18
53 PM](https://github.com/user-attachments/assets/4ea3416d-3930-4579-8337-94ead329189d)

Attempting to donate $4:
![Screenshot 2024-10-07 at 7 19
05 PM](https://github.com/user-attachments/assets/5b118072-7edb-43a6-aad8-ffee2e7a29e8)

# Implementation notes

Happy to tweak any formatting or anything else. 

I put a validation in the form for the $5 minimum but if someone tries
to get around that I also adjusted the minimum in
`MembershipPaymentForm`. This also changes the behavior for renewals
which isn't in the card but seems like the kind of thing we'd want
changed everywhere.