Try to use magick command and fallback to convert. (#1708)
# What it does

Resolves #1689 by attempting to use the `magick` command instead of
`convert`, which will be deprecated in Imagemagick 7. This deprecation
warning has started showing up on developer machines, and it's annoying.

For bonus fun, Heroku has an older version of imagemagick from before
there was a `magic` command. So I added a fallback for environments
where we only have the older command.