Remove @import statements from the SASS (#1716)
# What it does

A recent dependabot update for sass created more deprecation warnings,
so this clears that noise out. The biggest thing is that `@import` is
fully deprecated and `@use` is preferred now.

# Why it is important

`@import` will be completely removed in a future version of sass.

# Implementation notes

99% of the changes are the sass-migrator tool, 0.9% are my text editor
auto-formatting with prettier, and the last 0.1% is resolving a circular
dependency issue (`mixins.scss` and `globals.scss` depended on one
another).