Added renewal count to the renewal requests index page (#1740)
# What it does

Added the renewal count to the renewal requests index page and
rearranged a few of the columns.

# Why it is important

#1654 

# UI Change Screenshot

The renewal requests index page on a iPad mini
![Screenshot 2024-11-08 at 3 28
23 PM](https://github.com/user-attachments/assets/59fdb146-8fdf-45dd-a7a0-bb4e1ce1f5ba)

# Implementation notes

If this page needs to support a narrower screen size we can combine some
of the columns but the iPad mini (768px wide in portrait mode) seems
like it would cover most devices. Let me know if not!