Show the raw value of each item's status in addition to the member facing version in the admin interface (#1792)
# What it does

Displays `item.status` on the admin item's index and show pages in
addition to the `item_status_label` version.

# Why it is important

#1742 

# UI Change Screenshot

Admin items  index page:
![Screenshot 2024-12-17 at 4 33
21 PM](https://github.com/user-attachments/assets/9ee6750f-15e2-4927-bacc-31a46d572d22)

Admin item show page:
![Screenshot 2024-12-17 at 4 33
44 PM](https://github.com/user-attachments/assets/e3efb652-a48d-4067-823b-7068309e7a4c)

# Implementation notes

Just added the capitalized version of the status to both pages. Right
now it's background is always gray, but we could add some colors to make
it a little nicer if it's not too noisy.