Adds tagging to tickets on items (#1796)
# What it does

Allows admins to tag tickets and see them on the ticket index and show
pages.

# Why it is important

#1746 

# UI Change Screenshot

Tags on the index page:
![Screenshot 2024-12-18 at 3 35
29 PM](https://github.com/user-attachments/assets/eab9fcc3-331f-4ad4-bbea-71435ced7af2)

Tags on the show page:
![Screenshot 2024-12-18 at 3 37
17 PM](https://github.com/user-attachments/assets/f618af96-f482-4040-ae3a-823d05295e82)

Tags on the edit page: 
![Screenshot 2024-12-18 at 3 37
26 PM](https://github.com/user-attachments/assets/a187ebc9-a1a1-44db-9407-67a46553f7a4)

Tags on the new page:
![Screenshot 2024-12-18 at 3 37
41 PM](https://github.com/user-attachments/assets/92f99a96-4836-48c1-acbd-fb3ee995c741)

# Implementation notes

I rebased #1795 into this since the tagging gem requires `>= 3.2` Ruby.
Once that's merged this diff should be a little easier to understand.

Right now the tags just appear as light gray chips or pills. We could
add special styles for special tags (eg 'Urgent' with a red background)
but I think this is a good first pass.

The tags look case sensitive but actually aren't. So if you're the first
person to tag something as `Urgent` it will show up as `Urgent` event if
you change it to `urgent`. That also goes for any other tickets that are
tagged `Urgent`.

In terms of tests, I just added onto some existing controller tests.
Maybe system tests for this ticket stuff, but converting them felt like
a lot for right now. In the process of troubleshooting the gem setup I
added a model test that probably isn't _really_ necessary but that file
was empty anyway.