Validate that reservation changes will cover existing loans. (#1831)
# What it does

Adds a validation to close a gap in how we were validating
`ReservationHolds`. Currently, it's possible to reduce the `quantity` of
a `ReservationHold` beneath what's would be required by existing
`ReservationLoan` objects. So you could have 4 items added to a
reservation, but only 3 "reserved" via `ReservationHold` objects. This
breaks the lending model's guarantees, which are based around the
`ReservationHold` being the source of truth for item availability.

In the last meetup we had a good chat about other ways to handle this
situation, which I think are worth exploring as we iterate on the UI.
For now, though, I'd just like to maintain data integrity so we don't
end up promising more items than we can actually lend out.

Previously, only some `ReservationLoan` objects had a value for
`quantity` (those that were for items that aren't uniquely numbered). By
always having a value for that field, though, we can simply some of the
logic and do things like `reservation_loans.sum(:quantity)` to count how
many items are expected from the loans associated with a
`ReservationHold`.