Change link_to to button_to on exports page. (#1855)
# What it does

Fixes a bug that was preventing these buttons from working. I think they
broken during one of the Turbo upgrades- I think `link_to` with a method
doesn't work if Turbo is also disabled for the link.

Since this is an urgent matter, I'm going to merge and deploy this ASAP.