Admins can mark appointments as pulled in addition to pending and complete (#1860)
# What it does

When looking at the appointments index, admins will now see three groups
of appointments: pending, pulled, and completed. They can move
appointments between these three states and the page will update
accordingly.

They can also make these same changes on the appointment details/show
page.

# Why it is important

#1848

# UI Change Screenshot

Mobile version of the appointments index:
![Screenshot 2025-02-23 at 5 15
05 PM](https://github.com/user-attachments/assets/7d8ed7a7-b5ec-4d94-948a-0578ed1a9153)

Desktop version of the appointments index (the version before the
feature flag):
![Screenshot 2025-02-23 at 5 15
38 PM](https://github.com/user-attachments/assets/49a74ad9-210b-4496-be39-25f361667d70)

Desktop version of the appointments index (the version after the feature
flag):
![Screenshot 2025-02-23 at 5 15
21 PM](https://github.com/user-attachments/assets/d3649e25-2314-44dd-9e2b-aeef65ddbf79)

Show page for a pending appointment:
![Screenshot 2025-02-23 at 5 59
28 PM](https://github.com/user-attachments/assets/f8912faf-8ea4-4a54-abff-57e0f44c0cbd)

Show page for a pulled appointment:
![Screenshot 2025-02-23 at 5 58
06 PM](https://github.com/user-attachments/assets/4f586b5c-c3d4-4f8c-ad7c-000e5aa115ae)

Show page for a completed appointment:
![Screenshot 2025-02-23 at 5 58
18 PM](https://github.com/user-attachments/assets/24f31ae6-36c9-47c0-b8f6-26759dbc8b13)

# Implementation notes

I added a `pulled_at` column to the `appointment` since completion as a
concept is represented by a `completed_at` column. There is probably
some sort of refactoring that could be done around representing this
state, but I think there's a larger refactor that might be good.

There's a more than fair amount of duplication floating around here.
Specifically, there are mobile and desktop versions of appointment
partials, new and original versions of the index page and partials, and
a few controllers that I duped for this.

Before refactoring or drying anything up I think we need to find out
more about how librarians are using these pages. Do they always use the
old or the new version of the index page? Is there a better way to
represent the index page that's more useful and doesn't require desktop
and mobile specific markup?

That said, I think the verbosity of this is probably fine for the near
term and I'd rather be sure we dry things up the right way then guess
and end up painting ourselves into a corner. Plus, I think this will
definitely make things better for the librarians so I'm happy to put any
refactoring off to get this to them sooner.