Show maintenance tags (#1853)
# What it does
- Addresses #1804: Adds the ticket tags to the "items in maintenance"
report as well as a filter for the tags. This currently allows you to
filter on one tag at a time--if we need to filter by multiple tags at
once, we'll need to update this!

# Why it is important

Makes tags more helpful!


# UI Change Screenshot

<img width="1428" alt="Screenshot 2025-02-20 at 12 27 31 PM"
src="https://github.com/user-attachments/assets/93bff036-8388-429e-925e-9c17ef74e317"
/>


# Implementation notes

There is a strange bug where if you refresh the page occasionally, it'll
prepopulate old params into the filter fields (it will not filter the
results). This is annoying, but I thought it would not come up that
often. I think it has to do with the fact that I had to switch to a
non-ransack form helper because it was difficult to include the tags in
the ransack search (it should be compatible, but it that part of Ransack
seems stale--Ransack recently required that you have to allowlist
everything you search by explicitly, which I could not do easily with
tags).