Upgrade to Rails 8 (#1879)
# What it does

Upgrades Rails' Ruby and JavaScript dependencies to 8.0!

# Why it is important

#1814 and just general keeping up with the times.

# Implementation notes

There's a constant reassignment warning when running the tests that
happens when `load_tasks` is called in `test/tasks/holds_test.rb`. I
tried a few things but haven't cracked it yet. It's pretty minor so I
figure it shouldn't hold this up.

Aside from that, check out the specific comments I've left.