Fix system test
The assert_difference was not catching the database change from the
simulated UI actions. Changed strategies by asserting the expected UI
flash message shows up (which has the effect of causing the test to
appropriately wait for the request to finish) and then letting the
existing DB state check act as the main assertion.

Also removed a few comment lines from the test, as IMHO they were just
duplicating the lines of code.