Fix flakey unpulled items system tests and bin/setup errors (#1906)
# What it does

A lot of PRs were experiencing system test failures in
`test/system/admin/appointments_notifications_test.rb` that seems to be
a timing issue.

Less frequently but still a problem were PRs failing CI when running
`bin/setup` because a tool that requires approval would randomly be
selected to create a hold with.

This PR adds the [timecop gem](https://github.com/travisjeffery/timecop)
to prevent the timing issues in the test and it adds a scope for items
that excludes all tools that would require approval to borrow.

# Why it is important

Random failures in CI that aren't the fault of that PR are a huge
bummer, so this eliminates that bummer.

# Implementation notes

Just Timecop and the scope.