Add optional limit to how much of an item pool can be reserved by a single reservation (#1901)
# What it does

Fixes #1530 by allowing staff to set a percentage limit on how much of a
given item pool can be claimed by a single reservation.

# Why it is important

We don't want a single reservation to use the entire inventory of a
given item type.

# UI Change Screenshot

There is a new field on the ItemPool edit form to set this value:
![Screenshot 2025-04-04 at 4 26
56 PM](https://github.com/user-attachments/assets/7a27a731-ca2e-44f9-9579-1cd28f6be6ba)

When set, this limit is shown on the ItemPool show page
![Screenshot 2025-04-04 at 4 26
45 PM](https://github.com/user-attachments/assets/9b776566-eefd-45de-9da3-45ddf88839f5)

Here's an example of the validation working to limit how many items can
be reserved:
![Screenshot 2025-04-04 at 4 26
35 PM](https://github.com/user-attachments/assets/8c997190-5da3-4daf-bff4-b272b69c5de7)

# Implementation notes

* I thought about storing a number of items as an integer instead of a
percentage as specified in the ticket, but decided ultimately that a
percentage is better since it will automatically scale with the number
of items (so if we get more items, we don't have to go in and adjust
it). I think that's a feature, but it's worth confirming this with
staff.

---------

Co-authored-by: Michael Crismali <michael@crismali.com>