Link to borrow policy approvals on the member show page (#1916)
# What it does

Adds all of a member's borrow policy approvals and their current status
to the admin member's show page and links to the approval edit page.

# Why it is important

#1893 Makes borrow policy approvals (aka "C" tools) much easier for
librarians to administrate.

# UI Change Screenshot

![Screenshot 2025-04-27 at 4 56
40 PM](https://github.com/user-attachments/assets/aee6fdd7-1fec-4588-bb1c-76188ccb733a)

# Implementation notes

For now I just put them in the side panel but maybe they would be better
in one of the tabs on that page. I wanted to hold off on that because it
seems like a bigger push and I think we want to surface these to start
with. Maybe if we add something like changing their statuses from the
member page it'd make more sense to put them behind a tab.