Accessibility improvements for search (#1943)
# What it does

Makes the site a bit more accessible in a few places but mostly when
browsing items. See my comments on the code for more specifics. On the
pages I made changes the Lighthouse score in Chrome went from ~82 to ~90
(not that this score is the be all, end all measure of accessibility).

Also fixed an issue where the sort links would wrap weirdly on mobile:

![Screenshot 2025-06-09 at 2 57
05 PM](https://github.com/user-attachments/assets/8675fdd1-e1ca-401f-8624-f02da4d76604)

# Why it is important

Accessible sites are better everyone!

# UI Change Screenshot

Search results page on mobile (the wrapping is fixed for the sort links
but also the active sort is a slightly darker color than before):

![Screenshot 2025-06-09 at 2 51
25 PM](https://github.com/user-attachments/assets/0fb03c75-84a8-428a-af60-1c95dbd692ee)

Search results page on desktop:

![Screenshot 2025-06-09 at 2 51
43 PM](https://github.com/user-attachments/assets/53ca1450-5ce4-46f6-888a-d16a9e52aace)

# Implementation notes

There are a few other things Lighthouse flags as potential issues, but I
figured this was enough for one PR. Also, this is all based on my
interpretation of the [WAI-ARIA
standards](https://www.w3.org/WAI/standards-guidelines/aria/).