Cancel appointments for holds that are ended when an item is unavailable (#1955)
# What it does

<!-- Describe the changes that you've made to the application in this
PR. -->
- In the Hold model, adds a method to remove a hold from an appointment,
cancelling the appointment if there are no other items
- In the Item model, updates `clear_holds_if_inactive` to trigger holds
on the item to be removed from an appointment if it exists.

# Why it is important
Addresses #1934 
# UI Change Screenshot

<!-- If this PR makes a change to the UI put a screenshot here. If you
have before and after screenshots please add these. -->

# Implementation notes

- Thinking about this, it might make sense to add a notification here,
but wasn't sure if we wanted to avoid notifications for some reason!
- The way I implemented this ended up interacting badly with consumable
items, so that if you were the librarian checking out the last of a
consumable item, you'd see an error screen once the item was "loaned",
retiring the item once it was consumed (the scenario in the
consumable_concern_test I edited). I just fudged the test a little by
adding another item to avoid this situation in the test, since looking
at the description for the consumable borrow policy, we haven't had
consumable items since the seed share a few years ago. If we're still
using consumable items, I'm happy to look more into how to reconcile
these issues!

<!-- Anything notable about the technical approach you took. -->

<!-- Any open questions you have about the PR or areas where you'd like
specific feedback. -->