Add ability for member to view Documents agreed to (#1947)
# What it does

Resolves #874: feat: Add ability for a member to view the most recent
Document `'agreement'`, `'code_of_conduct'`, and `'borrow_policy'`.

# Why it is important

There is currently no way for members to view the policies they agree to
when they sign up for a membership.

# UI Change Screenshot

<img width="790" alt="Screenshot 2025-06-11 at 21 28 24"
src="https://github.com/user-attachments/assets/f86998f2-84bc-461d-afbd-43f30d173b28"
/>

# Implementation notes

There is no way to tie a particular Document to a Member agreeing, so
we're unable to provide a member with when they agreed to a particular
document (or a version of a document) for now, but that would be a good
enhancement in the future.

The view/controller method are very similar currently, but opted to
leave them distinct so that iteration on each version can be done
without affecting other versions.

---------

Co-authored-by: Jim Benton <jim@autonomousmachine.com>