Adds for later list items for members (#2020)
# What it does

This adds a saved for later page as well as buttons for adding/removing
items from a member's saved for later list.

I think this is a reasonable starting point for this feature and I'm
assuming it'll change a fair bit as we gather feedback before merging it
in (assuming we merge this in and don't just make a new PR, depending on
the feedback).

# Why it is important

I know some idea of this has been mentioned in stakeholder meetings
because some users will use holds to remind themselves of items they
need but don't otherwise have a way to save. So they make the hold and
then pass on the hold until they're ready to borrow it. This gives them
a place to save those items without messing with the loaning/holding
flow.

# UI Change Screenshot

I should note that users that are not signed in won't see any of these
buttons in the interface.

The for later page:
<img width="1004" height="966" alt="Screenshot 2025-11-01 at 5 16 14 PM"
src="https://github.com/user-attachments/assets/ac092e7f-1537-49cc-af50-762f1bef5138"
/>

The empty for later page:
<img width="988" height="259" alt="Screenshot 2025-11-01 at 5 15 15 PM"
src="https://github.com/user-attachments/assets/addd752f-7b88-4c71-8728-6e89d7fd08a8"
/>

The navigation:
<img width="256" height="272" alt="Screenshot 2025-11-01 at 5 22 59 PM"
src="https://github.com/user-attachments/assets/09fe5415-a281-436c-8298-4352e8323c15"
/>


Item details page when not wish listed:
<img width="1001" height="544" alt="Screenshot 2025-11-01 at 5 15 41 PM"
src="https://github.com/user-attachments/assets/8111d323-d6d6-41dc-8806-c2143e849327"
/>

Item details page when wish listed:
<img width="995" height="505" alt="Screenshot 2025-11-01 at 5 15 47 PM"
src="https://github.com/user-attachments/assets/47bd357c-55cf-4eab-af1f-f5ff830233d1"
/>

Items index page with a mix of wish listed and not wish listed items:
<img width="1033" height="806" alt="Screenshot 2025-11-01 at 5 15 29 PM"
src="https://github.com/user-attachments/assets/af15cfb6-7892-4e6d-8ab4-6d3951ae8ad4"
/>


# Implementation notes

~~I called it a "Wish List" but I don't think it would be so hard to
rename the concept. The public library refers to this feature as your
"For Later Shelf" but I wasn't sure that was right for this context. I
also thought about calling it "favorites" but to me that implied that it
would only be for tools you'd want to borrow over and over.~~

On the technical side, I implemented this with Turbo Streams so the
interface just automatically updates. If someone is browsing the site
without javascript turned on things should still work, they'll just be
redirected when they add or remove items.

In terms of styles I just use small buttons for adding/removing items.
The actual wish list page rips off a lot of markup from the index page.
I think a case could be made to make it a partial but imagine these
could end up diverging so I didn't want to start there.

**Update:** I've changed the model from `WishListItem` to
`ForLaterListItem`. Right now there isn't a `ForLaterList` model but I
think that would be simple enough to add later on (and convert the
current model into more of a join table). There are a lot of little UI
thinks to manage re multiple lists (when you click save is it a dropdown
of lists? a dialog? etc) so I think it's best to wait a bit on diving
into that.