Remove Organizations (#2058)
# What it does

As we re-target reservations toward camping and other special
collections for members, we won't need to have organizations signup.
This PR removes two database tables and a bunch of code related to orgs
that we aren't going to need.

Most of the reservation system will continue to be used, so it remains.

# Why it is important

No need to maintain code that isn't being used.